v.7.3.43

+++ app/src/main/java/com/example/projectnailsschedule/domain/models/ClientModelDb.kt
@@ -15,7 +15,7 @@ import kotlinx.parcelize.Parcelize
 @Entity(tableName = "clients")
 data class ClientModelDb(
     @PrimaryKey(autoGenerate = true)
-    val _id: Int? = null,
+    val _id: Long? = null,
 
     @ColumnInfo(name = "name")
     val name: String? = null,
